better handle 'no connection'

+++ doc/delete.js
@@ -28,6 +28,13 @@ module.exports = function(RED) {
             if (!U.keyHasValue(node, params, 'id')) return;
 
             const client = node.conn.client();
+            if (!client) {
+                node.warn("Not connected")
+                node.status({fill:"red",shape:"ring",text:"Not connected"})
+                U.slateStatusClear(node);
+                return
+            }
+
             node.status({fill:"blue",shape:"dot",text:"deleting"})
             client.delete(params).then(function (res) {
                 node.status({fill:"green",shape:"dot",text:res.result})
